requirements.txt
langgraph>=0.3.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
psycopg[binary]>=3.2.0
alembic>=1.13.0
SQLAlchemy>=2.0.0
streamlit>=1.40.0
pdfplumber>=0.11.0
python-docx>=1.1.0
requests>=2.32.0
beautifulsoup4>=4.12.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
numpy>=1.26.0
redis>=5.0.0
structlog>=24.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
nemoguardrails>=0.10.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
